# fix: Avoid serializing minidumps over procspawn