--- a/crates/symbolicator/Cargo.toml
+++ b/crates/symbolicator/Cargo.toml
@@ -11,7 +11,6 @@ anyhow = "1.0.38"
 apple-crash-report-parser = "0.4.2"
 backtrace = "0.3.55"
 base64 = "0.13.0"
-bytes = { version = "1.0.1", features = ["serde"] }
 cadence = "0.25.0"
 chrono = { version = "0.4.19", features = ["serde"] }
 console = "0.14.0"