--- a/crates/symbolicator/src/endpoints/applecrashreport.rs
+++ b/crates/symbolicator/src/endpoints/applecrashreport.rs
@@ -32,7 +32,11 @@ async fn handle_apple_crash_report_request(
 
         let content_disposition = field.content_disposition();
         match content_disposition.as_ref().and_then(|d| d.get_name()) {
-            Some("apple_crash_report") => report = Some(read_multipart_file(field).await?),
+            Some("apple_crash_report") => {
+                let mut report_file = tempfile::tempfile()?;
+                read_multipart_file(field, &mut report_file).await?;
+                report = Some(report_file)
+            }
             Some("sources") => sources = read_multipart_sources(field).await?.into(),
             Some("options") => options = read_multipart_request_options(field).await?,
             _ => (), // Always ignore unknown fields.