Instead of passing (and serializing) a byte buffer to procspawn, this will rather create a tempfile and pass that path via procspawn.

This seems to be a small win when processing a 20M minidump, though note that I was timing `cargo run -p process-event` so timings include overhead from both cargo and uploading of the minidump.

before:
  Time (mean ± σ):      1.611 s ±  0.041 s    [User: 274.8 ms, System: 102.0 ms]
  Range (min … max):    1.548 s …  1.697 s    10 runs
after:
  Time (mean ± σ):      1.526 s ±  0.024 s    [User: 275.7 ms, System: 101.5 ms]
  Range (min … max):    1.496 s …  1.573 s    10 runs

Due to some error handling and borrowcheck shenanigans, the minidumps are now persisted in the diagnostics cache for *all* errors during processing, while previously timeouts were ignored.