@flub 2021-07-30
> 2. Stream into a file directly in the endpoint, then pass around the file. @flub we've had various versions of this in the past already, do you remember where this was?

hmm, no i don't.  sorry