@Swatinem 2021-07-30
now streaming directly to a file (and updating rustc), I get these timings, though I noticed they have high variability:

before:
  Time (mean ± σ):      1.694 s ±  0.028 s    [User: 288.2 ms, System: 106.9 ms]
  Range (min … max):    1.656 s …  1.749 s    10 runs
after:
  Time (mean ± σ):      1.621 s ±  0.030 s    [User: 294.3 ms, System: 110.5 ms]
  Range (min … max):    1.585 s …  1.676 s    10 runs